+++ tests/test_jobs.py
@@ -17,7 +17,7 @@ import yaml
 from _pytest.logging import LogCaptureFixture
 from requests import HTTPError
 
-from webchanges import __project_name__ as project_name
+from webchanges import __project_name__
 from webchanges.config import CommandConfig
 from webchanges.handler import JobState
 from webchanges.jobs import BrowserJob, BrowserResponseError, JobBase, NotModifiedError, ShellJob, UrlJob
@@ -27,7 +27,7 @@ from webchanges.storage import CacheSQLite3Storage, Config, YamlConfigStorage, Y
 here = Path(__file__).parent
 data_path = here.joinpath('data')
 cache_file = ':memory:'
-cache_storage = CacheSQLite3Storage(cache_file)
+cache_storage = CacheSQLite3Storage(cache_file)  # type: ignore[arg-type]
 
 if sys.version_info[0:2] < (3, 8) and sys.platform == 'nt':
     # https://docs.python.org/3/library/asyncio-platforms.html#asyncio-windows-subprocess
@@ -140,6 +140,18 @@ TEST_ALL_URL_JOBS = [
 ]
 
 
+def new_command_config(config_file: Path, jobs_file: Path, hooks_file: Path) -> CommandConfig:
+    return CommandConfig(
+        args=[],
+        project_name=__project_name__,
+        config_path=here,
+        config_file=config_file,
+        jobs_def_file=jobs_file,
+        hooks_file=hooks_file,
+        cache=cache_file,  # type: ignore[arg-type]
+    )
+
+
 def test_kind() -> None:
     with pytest.raises(ValueError) as e:
         JobBase.unserialize({'kind': 'url'})
@@ -173,9 +185,6 @@ def test_run_job(
     event_loop: AbstractEventLoop,
 ) -> None:
     job = JobBase.unserialize(input_job)
-    if sys.version_info < (3, 8) and job.use_browser:
-        pytest.skip('Playwright testing requires Python 3.8')
-
     with JobState(cache_storage, job) as job_state:
         data, etag = job.retrieve(job_state)
         if job.filter == [{'pdf2text': {}}]:
@@ -212,9 +221,6 @@ def test_run_ftp_job_needs_bytes() -> None:
     ids=('BrowserJob' if v.get('use_browser') else 'UrlJob' for v in TEST_ALL_URL_JOBS),  # type: ignore[attr-defined]
 )
 def test_check_etag(job_data: Dict[str, Any], event_loop: AbstractEventLoop) -> None:
-    if sys.version_info < (3, 8) and job_data.get('use_browser'):
-        pytest.skip('Playwright testing requires Python 3.8')
-
     job_data['url'] = 'https://github.githubassets.com/images/search-key-slash.svg'
     job = JobBase.unserialize(job_data)
     with JobState(cache_storage, job) as job_state:
@@ -229,9 +235,6 @@ def test_check_etag(job_data: Dict[str, Any], event_loop: AbstractEventLoop) -> 
     ids=('BrowserJob' if v.get('use_browser') else 'UrlJob' for v in TEST_ALL_URL_JOBS),  # type: ignore[attr-defined]
 )
 def test_check_etag_304_request(job_data: Dict[str, Any], event_loop: AbstractEventLoop) -> None:
-    if sys.version_info < (3, 8) and job_data.get('use_browser'):
-        pytest.skip('Playwright testing requires Python 3.8')
-
     if job_data.get('use_browser'):
         pytest.skip('Capturing of 304 cannot be implemented in Chrome')  # last tested with Chromium 89
 
@@ -255,21 +258,16 @@ def test_check_etag_304_request(job_data: Dict[str, Any], event_loop: AbstractEv
     TEST_ALL_URL_JOBS,
     ids=('BrowserJob' if v.get('use_browser') else 'UrlJob' for v in TEST_ALL_URL_JOBS),  # type: ignore[attr-defined]
 )
-def test_check_ignore_connection_errors_and_bad_proxy(job_data: Dict[str, Any], event_loop: AbstractEventLoop) -> None:
-    if sys.version_info < (3, 8) and job_data.get('use_browser'):
-        pytest.skip('Playwright testing requires Python 3.8')
-
-    job_data['url'] = 'http://connectivitycheck.gstatic.com/generate_204'
-    job_data['http_proxy'] = 'http://notworking:ever@google.com:8080'
+def test_check_ignore_connection_errors(job_data: Dict[str, Any], event_loop: AbstractEventLoop) -> None:
+    job_data['url'] = 'http://localhost'
     job_data['timeout'] = 0.001
     job = JobBase.unserialize(job_data)
     with JobState(cache_storage, job) as job_state:
         job_state.process()
-        if job_state.exception and not isinstance(job_state.exception, BrowserResponseError):
-            assert sum(
-                list(x in str(job_state.exception.args) for x in ('Max retries exceeded', 'Timeout 1ms exceeded.'))
-            )
-        assert job_state.error_ignored is False
+        assert job_state.exception and any(
+            x in str(job_state.exception.args) for x in ('Max retries exceeded', 'Timeout 1ms exceeded.')
+        )
+        assert getattr(job_state, 'error_ignored', False) is False
 
     job_data['ignore_connection_errors'] = True
     job = JobBase.unserialize(job_data)
@@ -285,13 +283,34 @@ def test_check_ignore_connection_errors_and_bad_proxy(job_data: Dict[str, Any], 
     TEST_ALL_URL_JOBS,
     ids=('BrowserJob' if v.get('use_browser') else 'UrlJob' for v in TEST_ALL_URL_JOBS),  # type: ignore[attr-defined]
 )
+def test_check_bad_proxy(job_data: Dict[str, Any], event_loop: AbstractEventLoop) -> None:
+    job_data['url'] = 'http://connectivitycheck.gstatic.com/generate_204'
+    job_data['http_proxy'] = 'http://notworking:ever@localhost:8080'
+    job_data['timeout'] = 0.001
+    job = JobBase.unserialize(job_data)
+    with JobState(cache_storage, job) as job_state:
+        job_state.process()
+        if job_state.exception and not isinstance(job_state.exception, BrowserResponseError):
+            assert any(
+                list(
+                    x in str(job_state.exception.args)
+                    for x in ('Max retries exceeded', 'Read timed out', 'Timeout 1ms exceeded.')
+                )
+            )
+        assert job_state.error_ignored is False
+
+
+@connection_required  # type: ignore[misc]
+@pytest.mark.parametrize(  # type: ignore[arg-type,misc]
+    'job_data',
+    TEST_ALL_URL_JOBS,
+    ids=('BrowserJob' if v.get('use_browser') else 'UrlJob' for v in TEST_ALL_URL_JOBS),  # type: ignore[attr-defined]
+)
 def test_check_ignore_http_error_codes_and_error_message(
     job_data: Dict[str, Any], event_loop: AbstractEventLoop
 ) -> None:
     if job_data.get('use_browser'):
         pytest.skip('Cannot debug due to Playwrigth/Windows bug')  # TODO Remove this and fix
-    if sys.version_info < (3, 8) and job_data.get('use_browser'):
-        pytest.skip('Playwright testing requires Python 3.8')
 
     job_data['url'] = 'https://www.google.com/teapot'
     job_data['http_proxy'] = None
@@ -341,7 +360,7 @@ def test_stress_use_browser(event_loop: AbstractEventLoop) -> None:
 
         setup_logger()
 
-    urlwatch_config = CommandConfig([], project_name, here, config_file, jobs_file, hooks_file, cache_file)
+    urlwatch_config = new_command_config(config_file, jobs_file, hooks_file)
     urlwatcher = Urlwatch(urlwatch_config, config_storage, cache_storage, jobs_storage)
     urlwatcher.run_jobs()
 
@@ -499,7 +518,7 @@ def test_compared_versions() -> None:
     jobs_file = data_path.joinpath('jobs-time.yaml')
     config_storage = YamlConfigStorage(config_file)
     jobs_storage = YamlJobsStorage([jobs_file])
-    urlwatch_config = CommandConfig([], '', here, config_file, jobs_file, hooks_file, cache_file)
+    urlwatch_config = new_command_config(config_file, jobs_file, hooks_file)
     urlwatcher = Urlwatch(urlwatch_config, config_storage, cache_storage, jobs_storage)
 
     # compared_versions = 2