+++ webchanges/reporters.py
@@ -100,10 +100,7 @@ if os.name == 'nt':
     except ImportError as e:
         AnsiToWin32 = e.msg  # type: ignore[assignment,misc]
 
-try:
-    from zoneinfo import ZoneInfo  # not available in Python < 3.9
-except ImportError:
-    from backports.zoneinfo import ZoneInfo  # type: ignore[no-redef]
+from zoneinfo import ZoneInfo  # not available in Python < 3.9
 
 logger = logging.getLogger(__name__)