+++ CHANGELOG.rst
@@ -43,7 +43,8 @@ Unreleased
 
 Added
 -----
-* Support for Python 3.12.
+* Support for Python 3.12. Note that running in Python 3.12 requires the latest version of the ``greenlet`` package,
+  i.e ``greenlet>=3.0.0``.
 * ``data_as_json`` job directive for ``url`` jobs to indicate that ``data`` entered as a dict should be
   serialized as JSON instead of urlencoded and, if missing, the header ``Content-Type`` set to ``application/json``
   instead of ``application/x-www-form-urlencoded``.